Move git-sourcemaps-repository-data to helpers
