+++ src/commands/sourcemaps/upload.ts
@@ -6,13 +6,13 @@ import asyncPool from 'tiny-async-pool'
 import {URL} from 'url'
 
 import {ApiKeyValidator, newApiKeyValidator} from '../../helpers/apikey'
+import {getBaseSourcemapIntakeUrl} from '../../helpers/base-intake-url'
 import {InvalidConfigurationError} from '../../helpers/errors'
+import {getRepositoryData, newSimpleGit, RepositoryData} from '../../helpers/git-sourcemaps-repository-data'
 import {RequestBuilder} from '../../helpers/interfaces'
 import {getMetricsLogger, MetricsLogger} from '../../helpers/metrics'
 import {upload, UploadStatus} from '../../helpers/upload'
 import {getRequestBuilder} from '../../helpers/utils'
-import {getRepositoryData, newSimpleGit, RepositoryData} from './common/git-sourcemaps-repository-data'
-import {getBaseSourcemapIntakeUrl} from '../../helpers/base-intake-url'
 import {Sourcemap} from './interfaces'
 import {
   renderCommandInfo,
